issue 16: packages correction, probably due to merge
