+++ robo4j-hw-rpi/src/examples/java/com/robo4j/hw/rpi/i2c/adafruitbackpack/BiColor24BargraphExample.java
@@ -17,10 +17,6 @@
 
 package com.robo4j.hw.rpi.i2c.adafruitbackpack;
 
-import com.robo4j.hw.rpi.i2c.adafruitoled.BiColor;
-import com.robo4j.hw.rpi.i2c.adafruitoled.BiColor24BarDevice;
-import com.robo4j.hw.rpi.i2c.adafruitoled.PackElement;
-
 import java.util.concurrent.TimeUnit;
 
 /**
