+++ robo4j-hw-rpi/src/examples/java/com/robo4j/hw/rpi/i2c/adafruitbackpack/BiColor8x8MatrixFaceRotationExample.java
@@ -17,11 +17,6 @@
 
 package com.robo4j.hw.rpi.i2c.adafruitbackpack;
 
-import com.robo4j.hw.rpi.i2c.adafruitoled.BiColor;
-import com.robo4j.hw.rpi.i2c.adafruitoled.BiColor8x8MatrixDevice;
-import com.robo4j.hw.rpi.i2c.adafruitoled.LEDBackpackUtils;
-import com.robo4j.hw.rpi.i2c.adafruitoled.MatrixRotation;
-
 import java.io.IOException;
 import java.util.concurrent.TimeUnit;
