+++ robo4j-units-rpi/src/examples/java/com/robo4j/units/rpi/led/AdafruitBiColor24BackpackExample.java
@@ -20,9 +20,9 @@ package com.robo4j.units.rpi.led;
 import com.robo4j.RoboBuilder;
 import com.robo4j.RoboContext;
 import com.robo4j.RoboReference;
-import com.robo4j.hw.rpi.i2c.adafruitoled.BiColor;
-import com.robo4j.hw.rpi.i2c.adafruitoled.BiColor24BarDevice;
-import com.robo4j.hw.rpi.i2c.adafruitoled.PackElement;
+import com.robo4j.hw.rpi.i2c.adafruitbackpack.BiColor;
+import com.robo4j.hw.rpi.i2c.adafruitbackpack.BiColor24BarDevice;
+import com.robo4j.hw.rpi.i2c.adafruitbackpack.PackElement;
 
 import java.io.InputStream;
 import java.util.concurrent.Executors;
