+++ robo4j-units-rpi/src/main/java/com/robo4j/units/rpi/led/AbstractI2CBackpackUnit.java
@@ -20,10 +20,10 @@ package com.robo4j.units.rpi.led;
 import com.robo4j.ConfigurationException;
 import com.robo4j.RoboContext;
 import com.robo4j.RoboUnit;
-import com.robo4j.hw.rpi.i2c.adafruitoled.LEDBackpack;
-import com.robo4j.hw.rpi.i2c.adafruitoled.LEDBackpackFactory;
-import com.robo4j.hw.rpi.i2c.adafruitoled.LEDBackpackType;
-import com.robo4j.hw.rpi.i2c.adafruitoled.PackElement;
+import com.robo4j.hw.rpi.i2c.adafruitbackpack.LEDBackpack;
+import com.robo4j.hw.rpi.i2c.adafruitbackpack.LEDBackpackFactory;
+import com.robo4j.hw.rpi.i2c.adafruitbackpack.LEDBackpackType;
+import com.robo4j.hw.rpi.i2c.adafruitbackpack.PackElement;
 import com.robo4j.logging.SimpleLoggingUtil;
 import com.robo4j.units.rpi.I2CEndPoint;
 import com.robo4j.units.rpi.I2CRegistry;
