+++ robo4j-units-rpi/src/main/java/com/robo4j/units/rpi/led/Adafruit24BargraphUnit.java
@@ -20,10 +20,10 @@ package com.robo4j.units.rpi.led;
 import com.robo4j.ConfigurationException;
 import com.robo4j.RoboContext;
 import com.robo4j.configuration.Configuration;
-import com.robo4j.hw.rpi.i2c.adafruitoled.BiColor24BarDevice;
-import com.robo4j.hw.rpi.i2c.adafruitoled.LEDBackpack;
-import com.robo4j.hw.rpi.i2c.adafruitoled.LEDBackpackType;
-import com.robo4j.hw.rpi.i2c.adafruitoled.PackElement;
+import com.robo4j.hw.rpi.i2c.adafruitbackpack.BiColor24BarDevice;
+import com.robo4j.hw.rpi.i2c.adafruitbackpack.LEDBackpack;
+import com.robo4j.hw.rpi.i2c.adafruitbackpack.LEDBackpackType;
+import com.robo4j.hw.rpi.i2c.adafruitbackpack.PackElement;
 
 import java.util.List;
